# storage/api-docs/pos-api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "pDaftar POS Integration API",
        "description": "\n * Integratsiya API — pDaftar POS va uchinchi tomon kassalari (AliPOS, YesPOS va boshqalar) uchun.\n *\n * ## Asosiy qoida\n * pDaftar — **haqiqat manbai** (source of truth). Kassaning lokal bazasi faqat **kesh + navbat**.\n * Qoldiq har doim `stock_movements` ledgeridan hisoblanadi, kassa uni faqat ko'rsatadi.\n *\n * ## Offline ishlash\n * Har bir yozuv amali `client_operation_id` (UUID, kassada amal bajarilgan payt yaratiladi) bilan\n * yuboriladi. Takroriy yuborish hech narsani o'zgartirmaydi — server saqlangan javobni qaytaradi\n * (`replayed: true`). Internet paydo bo'lganda navbat `POST /sync/push` orqali paket qilib yuboriladi.\n *\n * ## Sotuv va qoldiq\n * Kassadan kelgan sotuv **rad etilmaydi**: tovar allaqachon xaridorga berilgan. Qoldiq yetmasa\n * minusga tushadi va bu ko'rinadigan muammo bo'lib qoladi — sinxronlashda yo'qolgan sotuvdan ko'ra yaxshiroq.\n *\n * ## Uchinchi tomon kassalari\n * Tashqi POS tokeni `pos:catalog.read`, `pos:sales.write`, `pos:clients.write` scope'lari bilan\n * beriladi — ular o'z omborini o'zi yuritadi, bizga faqat sotuvni yuboradi.\n * ",
        "contact": {
            "name": "pDaftar",
            "url": "https://pdaftar.uz"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8083",
            "description": "Lokal (Docker)"
        },
        {
...
        }
...
    "paths": {
        "/api/pos/v1/sales": {
            "post": {
                "tags": [
                    "POS Sales"
                ],
                "summary": "Sotuv qilish",
                "description": "\n     * pDaftardagi Sotuv bilan bir xil yozuvlarni yaratadi: frozen SalesCalcList + qatorlar → Debt →\n     * `stock_movements` (sale) → to'langan bo'lsa Repayment → Kassa Kirim. Ya'ni POS sotuvi va ilova\n     * sotuvi bazada bir xil ko'rinadi.\n     *\n     * Qoldiq yetmasa ham **rad etilmaydi** — minusga tushadi.\n     * ",
                "operationId": "4e39ed381af598727e19db3347c6c87d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PosSaleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sotuv yozildi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PosOperationResult"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Takroriy yuborish — avvalgi javob qaytarildi (replayed=true)"
                    },
                    "403": {
                        "description": "Scope yetarli emas"
                    },
                    "409": {
                        "description": "client_operation_id boshqa ma'lumot bilan ishlatilgan"
                    },
                    "422": {
                        "description": "Validatsiya xatosi"
                    }
                },
                "security": [
                    {
                        "posToken": []
                    }
                ]
            }
        },
        "/api/pos/v1/sales/cancel": {
            "post": {
                "tags": [
                    "POS Sales"
                ],
                "summary": "Sotuvni bekor qilish",
                "description": "Qoldiqni tiklaydi — sotuv harakatlari ledgerdan o'chiriladi, kompensatsiya yozuvi yozilmaydi, shuning uchun mahsulot tarixi 'sotuv bo'lmagan' ko'rinishida qoladi.",
                "operationId": "cc69c665567c70f0ef24b418761af5cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_operation_id",
                                    "sale_id"
                                ],
                                "properties": {
                                    "client_operation_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "sale_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bekor qilindi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PosOperationResult"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "posToken": []
                    }
                ]
            }
        },
        "/api/pos/v1/products": {
            "post": {
                "tags": [
                    "POS Write"
                ],
                "summary": "Mahsulot yaratish",
                "description": "Barcode bo'yicha mavjud mahsulot topilsa — yangi qator yaratilmaydi, mavjudi yangilanadi (qoldiq ikkiga bo'linib ketmasligi uchun). `quantity` yuborilsa `opening` harakati sifatida yoziladi.",
                "operationId": "46106ff96c88d797e7caebda7e6790d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_operation_id",
                                    "name"
                                ],
                                "properties": {
                                    "client_operation_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "local_id": {
                                        "description": "Offline yaratilgan ID. Bir paket ichida keyingi amallar bu mahsulotga local-havola obyekti orqali murojaat qiladi — /sync/push tavsifiga qarang.",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "barcode": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "quantity": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "unit_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "currency_id": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Yaratildi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PosOperationResult"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "posToken": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "POS Write"
                ],
                "summary": "Mahsulotni tahrirlash",
                "description": "Yuborilmagan maydonlar tegilmaydi (array_key_exists), aniq `null` esa tozalaydi.",
                "operationId": "8a5b1f9bc178ca6490fb8f9455543a35",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_operation_id",
                                    "id"
                                ],
                                "properties": {
                                    "client_operation_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "barcode": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Yangilandi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PosOperationResult"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "posToken": []
                    }
                ]
            }
        },
        "/api/pos/v1/clients": {
            "post": {
                "tags": [
                    "POS Write"
                ],
                "summary": "Mijoz qo'shish",
                "operationId": "bb6b45c36ff1f71a2116bf2b2cefd566",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_operation_id",
                                    "name"
                                ],
                                "properties": {
                                    "client_operation_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "address": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Yaratildi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PosOperationResult"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "posToken": []
                    }
                ]
            }
        },
        "/api/pos/v1/suppliers": {
            "post": {
                "tags": [
                    "POS Write"
                ],
                "summary": "Ta'minotchi yaratish",
                "operationId": "72bad09140a1133d95983701786942fc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_operation_id",
                                    "name"
                                ],
                                "properties": {
                                    "client_operation_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Yaratildi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PosOperationResult"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "posToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "PosOperationResult": {
                "title": "Amal natijasi",
                "properties": {
                    "client_operation_id": {
                        "description": "OpenAPI root for the POS Integration API.\n\nThis class holds no logic — it exists so the document's Info, servers,\nsecurity scheme and shared schemas live in one place instead of being\nscattered across whichever controller happened to be written first.",
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "type": "string",
                        "example": "sale.create"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "applied",
                            "failed"
                        ]
                    },
                    "replayed": {
                        "description": "true — bu amal avval bajarilgan, hech narsa o'zgarmadi. Kassa uni navbatdan o'chiradi.",
                        "type": "boolean"
                    },
                    "data": {
                        "type": "object"
                    },
                    "error": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "posToken": {
                "type": "http",
                "description": "Kassa tokeni — /terminals/register javobidan olinadi. Ro'yxatdan o'tish uchun oddiy pDaftar user tokeni ishlatiladi.",
                "scheme": "bearer"
            }
        }
    },
...
}